version 1.1.0
This version is significant because it includes a major change to how
values are set and changes are detected between the select box and the
Wui field. Previous to this version, if jQuery was used to set a value
directly on the select element, selecting that element from the combo
would not work because the selectedIndex of the select was not updated.
These flaws are fixed as well as some other minor improvements
including a CSS update in IE8, and the Wui.isset() function.