fixed xss vulnerability in highlighting code Updated highlight code to safely encode original text being highlighted (#1)
To reproduce: create an item in the combo box with the literal text
<img src=x onerror=prompt("XSS”);> inside of a span node in the li
stored in the combo box. Something like this:
<li class="wui-list-item wui-selected"><span
        class="ts-wui-company-name"> <span
        class="wui-highlight">djfdjkfjdksjfkjdsk</span><span
        class="wui-highlight">djfdjkfjdksjfkjdsk</span></span><span
        class="ts-wui-type">business<span
        class="wui-highlight">djfdjkfjdksjfkjdsk</span></span><span
        class="ts-wui-admin">#"&gt;&lt;img src=x
onerror=prompt("XSS");&gt;<span
            class="wui-highlight">djfdjkfjdksjfkjdsk</span></span><span
        class="ts-wui-url">#"&gt;&lt;img src=x
onerror=prompt("XSS");&gt;<span
        class="wui-highlight">djfdjkfjdksjfkjdsk</span></span></li>

Use a search term that is in one of the child nodes, but not in the
span with the xss attack. When it attempts to highlight the term in the
child spans it would run the script.